create module for encoder
