+++ src/fast_forward/index/disk.py
@@ -15,7 +15,7 @@ if TYPE_CHECKING:
     from collections.abc import Iterable, Iterator
     from pathlib import Path
 
-    from fast_forward.encoder import Encoder
+    from fast_forward.encoder.base import Encoder
 
 LOGGER = logging.getLogger(__name__)
