+++ src/fast_forward/index/memory.py
@@ -10,7 +10,7 @@ from fast_forward.index.base import IDSequence, Index, Mode
 if TYPE_CHECKING:
     from collections.abc import Iterable, Iterator
 
-    from fast_forward.encoder import Encoder
+    from fast_forward.encoder.base import Encoder
     from fast_forward.quantizer import Quantizer
 
 LOGGER = logging.getLogger(__name__)
