+++ src/fast_forward/util/indexer.py
@@ -9,7 +9,7 @@ from tqdm import tqdm
 if TYPE_CHECKING:
     from collections.abc import Iterable, Sequence
 
-    from fast_forward.encoder import Encoder
+    from fast_forward.encoder.base import Encoder
     from fast_forward.index.base import IDSequence, Index
     from fast_forward.quantizer import Quantizer
 
@@ -17,7 +17,10 @@ LOGGER = logging.getLogger(__name__)
 
 
 class IndexingDict(TypedDict):
-    """Dictionary that represents a document/passage, consumed by `Indexer`."""
+    """Dictionary that represents a document/passage.
+
+    Consumed by `Indexer.from_dicts`.
+    """
 
     text: str
     doc_id: str | None
